+++ src/main/java/io/hhplus/tdd/point/domain/entity/UserPoint.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.domain.entity;
 
 public record UserPoint(
         long id,
