+++ src/main/java/io/hhplus/tdd/point/domain/service/PointService.java
@@ -1,6 +1,7 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.domain.service;
 
 import io.hhplus.tdd.database.UserPointTable;
+import io.hhplus.tdd.point.domain.entity.UserPoint;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
