+++ src/test/java/io/hhplus/tdd/database/PointHistoryTableTest.java
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.database;
 
-import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.TransactionType;
+import io.hhplus.tdd.point.domain.entity.PointHistory;
+import io.hhplus.tdd.point.domain.vo.TransactionType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
