+++ src/test/java/io/hhplus/tdd/point/PointHistoryServiceTest.java
@@ -1,6 +1,9 @@
 package io.hhplus.tdd.point;
 
 import io.hhplus.tdd.database.PointHistoryTable;
+import io.hhplus.tdd.point.domain.entity.PointHistory;
+import io.hhplus.tdd.point.domain.service.PointHistoryService;
+import io.hhplus.tdd.point.domain.vo.TransactionType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
